Split out and move Risc0RequestEvaluator to risc0 module
