+++ crates/boundless-market/src/lib.rs
@@ -125,3 +125,7 @@ pub use storage::{
 /// Utility functions and types used elsewhere.
 pub(crate) mod util;
 pub use util::NotProvided;
+
+/// risc0 specific objects
+#[cfg(not(target_os = "zkvm"))]
+pub mod risc0;
