+++ crates/boundless-market/src/prover_utils/mod.rs
@@ -31,7 +31,6 @@ pub use request_evaluator::{
     EvaluationLimits, EvaluationMetrics, EvaluationRequest, ImageUploadCache, ImageUploadCacheKey,
     InputCacheKey, NativeWork, NormalizedWork, PreflightCache, PreflightCacheKey,
     PreflightCacheValue, PriorityRequestorCheck, RequestEvaluation, RequestEvaluator,
-    Risc0Evaluator,
 };
 
 use crate::{
